pyproject.toml
[tool.poetry]
name = "license-plate-recognition"
version = "0.1.0"
description = ""
authors = ["Dawid Badurek <dawid.badurek@besidethepark.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.10.0.84"
pillow = "^11.0.0"
ipykernel = "^6.29.5"
torch = "^2.5.1"
torchvision = "^0.20.1"
torchaudio = "^2.5.1"
matplotlib = "^3.9.3"
scikit-learn = "^1.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
setup-darknet = "scripts.setup_darknet:main"